Time: 1 ms (99.77%), Space: 40.9 MB (6.16%) - LeetHub

+++ 0441-arranging-coins/0441-arranging-coins.java
@@ -1,5 +1,5 @@
 class Solution {
     public int arrangeCoins(int n) {
-        return (int)(-1 + Math.sqrt(8*n+1))/2;
+        return (int) ((-1 + Math.sqrt(8 * (long) n + 1)) / 2);
     }
 }
